salesManagerからJSON読み書き機能をjsonManagerに抽出
